requirements.txt
streamlit==1.28.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
Pillow==10.0.1